pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.124.4",
    "jinja2>=3.1.6",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.21",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.38.0",
]
